feat: add map generation (hidden)
Add the possibility to retrieve a map with all the current velox
displayed as markers. Unfortunately, google maps doesn't allow to
generate static maps for free (static maps APIs are currently priced at
0.002 USD per map).

For this reason, the trick of using direction apis is used.
While this works ok on desktop, it doesn't produce the wanted result on
mobile.

Therefore, this functionality is currently hidden (i.e. BotFather
doesn't list the command), but it can be manually used via the
"/show_map" command.